+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-CustomTemplates-client/src/test/java/com/graphql_java_generator/samples/basic/client/ValidateCustomQueryIT.java
@@ -1,27 +1,25 @@
 package com.graphql_java_generator.samples.basic.client;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import com.graphql_java_generator.exception.GraphQLRequestExecutionException;
 import com.graphql_java_generator.exception.GraphQLRequestPreparationException;
 import com.graphql_java_generator.samples.customtemplates.client.graphql.forum.client.Board;
 import com.graphql_java_generator.samples.customtemplates.client.graphql.forum.client.QueryType;
 
-
 @SpringBootTest()
-@RunWith(SpringRunner.class)
-@ContextConfiguration(
-		classes = ValidateCustomQueryConfiguration.class)
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes = ValidateCustomQueryConfiguration.class)
 class ValidateCustomQueryIT {
 
 	@Autowired
@@ -34,5 +32,5 @@ class ValidateCustomQueryIT {
 		assertTrue(response.size() > 0);
 		assertTrue(response.get(0) instanceof Board);
 	}
-	
+
 }